Move certs helper next to call sites
