+++ tests/unit/cmk/utils/crypto/test_certificate.py
@@ -15,7 +15,7 @@ import pytest
 import time_machine
 from dateutil.relativedelta import relativedelta
 
-from tests.testlib.certs import rsa_private_keys_equal
+from tests.unit.cmk.utils.crypto.certs import rsa_private_keys_equal
 
 from cmk.utils.crypto.certificate import (
     Certificate,
