+++ tests/unit/cmk/utils/test_certs.py
@@ -13,7 +13,7 @@ import cryptography.x509 as x509
 import time_machine
 from dateutil.relativedelta import relativedelta
 
-from tests.testlib.certs import rsa_private_keys_equal
+from tests.unit.cmk.utils.crypto.certs import rsa_private_keys_equal
 
 from livestatus import SiteId
